ragX12-api/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
faiss-cpu
openai
